test: consolidate redundant OIDC end-to-end test classes (#2030)
The two end-to-end OIDC front-channel test classes existed only to exercise the
now-removed ForceOidc toggle (true vs false). With the toggle gone they run in the
same config, so EndToEnd_OidcFrontChannelControllerTest (the former ForceOidc=false
class) was a redundant subset: its two tests (TC1, TC2) duplicate the other class
with strictly weaker assertions (bare logout-URL prefix + event count, vs the full
federated logout URL and downstream-logout verification).

- Delete the redundant EndToEnd_OidcFrontChannelControllerTest (weaker TC1/TC2).
- Rename EndToEndForceOidc_OidcFrontChannelControllerTest -> EndToEnd_OidcFrontChannelControllerTest
  (the ForceOidc distinction no longer exists) and refresh its class doc.

The kept class is the superset (14 tests) with the stronger, current-behaviour
assertions. All 14 pass.
